user entity file renamed user.entity
